+++ lib/ConfigReader.ml
@@ -64,11 +64,18 @@ let createDisallowEmbeddedRegexLiteralRule options =
   end) in
   (module M : Rule.HASRULE)
 
-module DisallowedDeadCodeRuleModule = DisallowedDeadCodeRule.Make (struct
-  type options = DisallowedDeadCodeRule.Options.options
+let createDisallowAttributeRule options =
+  let open Yojson.Basic.Util in
+  let attribute = options |> member "attribute" |> to_string in
+  let suggestion =
+    match options |> member "suggestion" with `Null -> None | json -> Some (to_string json)
+  in
+  let module M = DisallowedAttributeRule.Make (struct
+    type options = DisallowedAttributeRule.Options.options
 
-  let options = DisallowedDeadCodeRule.default
-end)
+    let options = {DisallowedAttributeRule.Options.suggestion; DisallowedAttributeRule.Options.attribute}
+  end) in
+  (module M : Rule.HASRULE)
 
 let parseConfig path =
   let json = Yojson.Basic.from_file path in
@@ -91,7 +98,9 @@ let parseConfig path =
     | "DisallowEmbeddedRegexLiteral" ->
         let options = json |> member "options" in
         createDisallowEmbeddedRegexLiteralRule options
-    | "DisallowDeadCode" -> (module DisallowedDeadCodeRuleModule)
+    | "DisallowAttribute" ->
+        let options = json |> member "options" in
+        createDisallowAttributeRule options
     | _ -> raise RuleDoesNotExist
   in
   json |> member "rules" |> to_list |> List.map filter_rule